Factor out helper classes for dynamic docsum field writer.
